requirements.testing.txt
mock==1.0.1
python-ldap-test==0.2.0
